requirements.txt
Django==5.0
djangorestframework==3.14.0
django-filter==23.5
docker==7.0.0
aiohttp==3.9.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
